Move current swap components into subfolder
