+++ src/screens/Swap/FormOrHistory/Form/Error.js
@@ -5,10 +5,11 @@ import { StyleSheet, View } from "react-native";
 import { useNavigation, useRoute, useTheme } from "@react-navigation/native";
 import SafeAreaView from "react-native-safe-area-view";
 import { Trans } from "react-i18next";
-import GenericErrorView from "../../../components/GenericErrorView";
-import Button from "../../../components/Button";
-import { ScreenName } from "../../../const";
-import { TrackScreen } from "../../../analytics";
+
+import GenericErrorView from "../../../../components/GenericErrorView";
+import Button from "../../../../components/Button";
+import { ScreenName } from "../../../../const";
+import { TrackScreen } from "../../../../analytics";
 
 const forceInset = { bottom: "always" };
