+++ src/screens/Swap/FormOrHistory/History/OperationRow.js
@@ -2,16 +2,18 @@
 import React, { useCallback } from "react";
 import { TouchableOpacity, StyleSheet, View } from "react-native";
 import { useNavigation, useTheme } from "@react-navigation/native";
+import Icon from "react-native-vector-icons/dist/Ionicons";
+
+import type { MappedSwapOperation } from "@ledgerhq/live-common/lib/exchange/swap/types";
 import {
   getAccountUnit,
   getAccountName,
 } from "@ledgerhq/live-common/lib/account/helpers";
-import type { MappedSwapOperation } from "@ledgerhq/live-common/lib/exchange/swap/types";
-import Icon from "react-native-vector-icons/dist/Ionicons";
 
-import LText from "../../../components/LText";
-import CurrencyUnitValue from "../../../components/CurrencyUnitValue";
-import { ScreenName } from "../../../const";
+import LText from "../../../../components/LText";
+import CurrencyUnitValue from "../../../../components/CurrencyUnitValue";
+import { ScreenName } from "../../../../const";
+
 import SwapStatusIndicator from "../SwapStatusIndicator";
 
 const OperationRow = ({ item }: { item: MappedSwapOperation }) => {
