+++ src/screens/Swap/FormOrHistory/NotAvailable.js
@@ -5,8 +5,9 @@ import SafeAreaView from "react-native-safe-area-view";
 import { StyleSheet, View } from "react-native";
 import Icon from "react-native-vector-icons/dist/AntDesign";
 import { useTheme } from "@react-navigation/native";
-import LText from "../../components/LText";
-import { TrackScreen } from "../../analytics";
+
+import LText from "../../../components/LText";
+import { TrackScreen } from "../../../analytics";
 
 const NotAvailable = () => {
   const { colors } = useTheme();
