+++ src/screens/Swap/FormOrHistory/SwapStatusIndicator.js
@@ -5,8 +5,9 @@ import { View, StyleSheet } from "react-native";
 import IconAD from "react-native-vector-icons/dist/AntDesign";
 import { operationStatusList } from "@ledgerhq/live-common/lib/exchange/swap";
 import { useTheme } from "@react-navigation/native";
-import IconSwap from "../../icons/Swap";
-import { rgba } from "../../colors";
+
+import IconSwap from "../../../icons/Swap";
+import { rgba } from "../../../colors";
 
 export const getStatusColor = (
   status: string,
